Fix flickering music control widget under certain conditions.
In case the active audio player abandons audio focus while the playback
state is still set to 'playing', the transport control can start
flickering in the keyguard.
In that case, the transport control is removed from the keyguard host
because it has no listener anymore, but later is re-added due to
isMusicPlaying() still returning true. Re-adding it leads to it
re-registering with audio service, which clears the display due to
nothing being connected, which leads to the detachment listener being
called again, which in turn closes the loop.

While this may be a sign of a buggy player app (it should stop its
playback prior to abandoning audio focus), this situation still should
be handled gracefully. Fix the problem by simply reporting no music
being played while no remote control client is connected.

Change-Id: Id6834d38bc6faa3b6455ac66b413443f85606785