SystemUI: Set Status bar opaque during drawer pulldown on tablets only
The status bar only needs to be set as opaque when the notification drawer is pulled down on 600dp and 720dp devices. On phones, this is unnecessary as the top header of the drawer covers the status bar on the phone UI. This change makes a huge difference in reponsiveness on devices that run the phone UI as there is no longer a delay before the drawer appears when interacting with it on an activity with transclucent window decor.